+++ src/components/GroupCreate/GroupCreateNickNameSection.tsx
